backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.10
pydantic-settings==2.11.0
python-dotenv==1.1.1
pytest==8.4.2
httpx==0.28.1
PyJWT==2.10.1
passlib==1.7.4
